third_party/WebKit/LayoutTests/fast/dom/Range/range-isPointInRange-expected.txt

CONSOLE WARNING: line 65: 'Range.detach' is now a no-op, as per DOM (http://dom.spec.whatwg.org/#dom-range-detach).
test 1 passed

test 2 passed

test 3 passed

test 4 passed

test 5 passed

test 6 passed

test 7 passed

test 8 passed

test 9 passed

test 10 passed
